changing x-y coordinate incrementing function to to scan horizontally rather than diagonally
The last commit seems to have been counterproductive.  I interpreted the
fillit instructions to mean that subsequent guesses should be placed in
the left-upper most coordinates of the board.

By changing the incrementing function, we change the order in which solutions in the
search tree are visited.

42FileChecker assumes that
this:
is more "upper-left"
	than this
....#    .....
.....    #....
.....    .....
.....    .....
.....    .....

so, the incrementing function now follows that convention.